Migrating action creators to Redux saga (#77)
This patch leverages redux-saga and moved the action creators logic to dedicated sagas. 

Actions are now just simple synchronous descriptors, and sagas handle the asynchronous 
request stuff, triggering actions to notify the reducers when they're processed.

That makes the codebase better organized and easier to read especially because of the use of 
generators.