Bump v1.0.0. (#137)
### Release notes

This is the first major stable version of kinto-admin. Since latest v0.16.1, the codebase has been entirely rewritten to switch from an offline-first approach to a simpler live HTTP one. We really wish we could have introduced such a breaking change more gradually, but shifting that heavily can hardly be progressive to be honest. 

Now, the admin is more powerful and feature-complete than ever; here are a few highlights: 

- You don't need to provide a configuration file anymore as resources can be discovered and browsed directly from the UI;
- You can now create buckets and collections from the admin, configure their metadata (eg. collection schemas) and even delete them;
- We've added support for [Firefox Account](https://accounts.firefox.com/) authentication along the *BasicAuth* one; we're planning on introducing more alternative OAuth2 auth providers in the future;
- We've added native support for [kinto-attachments](https://kinto.github.io/kinto-attachment/), which means you can now attach files to records and post them to the server if the capability is available;

We really hope you'll enjoy this new version as much as we've loved building it. Of course there are and will be bugs, but you already know [where to file them](https://github.com/Kinto/kinto-admin/issues) :)