Notifications: Separate out a `StateProps`
This has to be an exact object and corresponds to what is returned by
mapStateToProps.

We also have to export the `Props` type, which is the type of the
"connected" props, with state and dispatch functions connected up by
the container.

The functions in `Props` correspond to action
dispatchers (a.k.a. "action creators"). Previously, we explicitly
named the actions that the component could dispatch, which is kind of
cool. Let's keep that, but explicitly re-use the notification actions
from the `actions/Notifications.js`, instead of retyping their type
signature.

Previously, these actions were typed as returning void, but actually
they return the same action that's being dispatched. We ignore the
returned action so just use "*" where needed so that it will
typecheck.