App: update typings
This one is a bit trickier than Notifications because it has
OwnProps (from where it's called in routes.js).

The OwnProps must be an exact object, like StateProps and
DispatchProps, so we can't pass extra props. The props given by Route
includes match, location, and history, none of which are used by App.