CollectionRecords: update typings
In order to make the typings line up, I re-used the signature for
deleteRecord. Unfortunately, we seem not to always have an `rid`
present, so special case that. We'll have to revisit.

Additionally, not all calls to `deleteRecord` pass a `last_modified`,
so make it optional. In the underlying kinto-http layer, it is, but
it shouldn't be being called in situations where it's missing, so
we'll have to revisit this too.

I also understood that flow has some confusing behaviors with
intersections of object types, so just convert stuff to exact objects
and combine them using spread. I'm not really sure why this works but
it seems to.

This also corrects a bug in the typings that isn't actually a real bug
-- deleteRecord and redirectTo are missing from the props passed to
the component (although they're actually being passed).