Introduce "update" types and use them everywhere
The data we provide to "update" functions includes at least one of
data and permissions but not necessarily both. This makes it unlike
the data you get back for those resources. Formalize this in the types
of some actions and use them to fix the Attributes and Permissions
components that use them.

While we're here, tighten up some of the Record containers
types (passing more actions than we need again).