AuthForm: update typings
In order to get the StateProps to line up, I had to correct an error
where the history field is declared to be both an array of strings and
an array of history entries. I'm not a Redux expert but it seems like
the history field is managed by the history reducer, which maintains a
state which is an array of history entries, so this means that this
probably wasn't a real error.

Despite getting errors that the StateProps were wrong, I didn't get
any errors complaining that the history reducer state was incompatible
with the app state. I suspect the typing isn't perfect yet, which is a
little scary, but that's JavaScript.