routes: update typings
This module contains wrappers around react-router stuff that tries to
extract breadcrumb information and fire route-updated
messages. Because react-router is higher-order, it's a bit
trickier. The solution this PR pushes towards isn't really complete or
the cleanest, but it gets us to a passing build and I'm running out of
time to work on this.

`routeCreator` is a wrapper around `Route` that is also a connected
component, so we have to define its prop types. I don't see any
obvious way to get it to automatically infer the props that `Route`
uses and the type is not exported so we redefine it here, more or
less.

`Route` does not take `title` so be careful not to pass it in the
props that go to `Route`.

`Route` is a higher-order-component that wraps our component (using
`render`). Our components are all connected components so they don't
get any particular props, but they all get the React Router
props (match, location, and history) which are defined in
`ContextRouter`.

We also insert a `ComponentWrapper` which wraps our component. This is
what we actually give to `Route`. It gets all the same router props
plus a component to call and the connected onRouteUpdated function.