Provide object to destructure if error data is null
In getErrorDetails, we attempted to destructure a `ClientError`'s
`data` property. But since that property is possibly undefined, we would
get a type error.

This commit fixes this error by providing an empty object as a
fallback in case the error's data property is undefined. It also updates a saga
test where we call this funtion. We were previously mocking `notifyError`,
but this is where the bug bubbled up. Instead, we now remove that mock
and use the real function in the saga test to prevent a regression.