Migrate away from container components, use `react-redux` hooks (#2378)
In this PR, we start the process of removing "container" components from
throughout the app where we `map{State, Disptach}ToProps`. Instead, we'll use
[react-redux hooks](https://react-redux.js.org/api/hooks) to provide dispatch
and Redux store values to components. This will lead to an overall fewer number
of components to manage. We started the process with Layout and Sidebar 
because they were relatively simple components to migrate.

This PR adds a typed useSelector hook -- a typed useDispatch hook will have to
come later because the way we set up the initial store is a little atypical.

Also, in order to use functional components, we had to update tests to allow for
new ways of providing the expected Redux state.