Switch App to a functional component, move session setup
In an effort to convert class based components to functional components
with hooks, we converted the App component. The tricky part with this
component was that it used an unconventional method of setting up the
redux store in the class based component's constructor method. This made
it convenient since in the same place we set up the Redux store, we
could also fire an initial saga that set up the session. But now that we
use a functional component, we no longer had access to the store or
dispatch through hooks.

Since the initial setup only affects the presentaiton of the HomePage
component though, we were able to move the that initial setup logic
there. Tests also moved the the HomePage_test file.