Refactor redux setup, migrate App to functional component (#2426)
* Install, switch to Redux Toolkit for configuring Redux store

We're not going all-in with Redux Toolkit patterns yet, but this commit does
use it for its `combineReducers` and new `configureStore` API to set up
middleware (the core `redux` `createStore` method was recently marked as
deprecated).

* Modify redux typing
- Use `ReturnType` of `store.getState`
- Add `AppDispatch` type
- Add `useAppDispatch` hook

* Switch App to a functional component, move session setup

In an effort to convert class based components to functional components
with hooks, we converted the App component. The tricky part with this
component was that it used an unconventional method of setting up the
redux store in the class based component's constructor method. This made
it convenient since in the same place we set up the Redux store, we
could also fire an initial saga that set up the session. But now that we
use a functional component, we no longer had access to the store or
dispatch through hooks.

Since the initial setup only affects the presentation of the HomePage
component though, we were able to move the that initial setup logic
there. Tests also moved the the HomePage_test file.

* Upgrade react-redux

* Downgrade @types/history to 4.x version

matches the 4.x version of `history` that we're pinned to